requirements.txt
dearpygui
imageio
imageio-ffmpeg
lpips
matplotlib
ninja
opencv-python
packaging
pandas
PyMCubes
pysdf
rich
scipy
tensorboardX
torch-ema
torchmetrics
tqdm
trimesh
scikit-image
kornia